feat(nonScalingSize-plugin): replace 'keepAspect' plugin with 'nonScalingSize' plugin
The 'keepAspect' plugin has been deprecated and replaced with the 'nonScalingSize' plugin. The code and test cases have been updated to reflect this change. The new 'nonScalingSize' plugin maintains the size of SVG elements when resized. For maintains only the stroke is necessary to use the standard attribute 'vector-effect="non-scaling-stroke"'.