refactor(plugin-animateTo): refactor animateTo and improve animation functionality
The animateTo has been significantly refactored to streamline the code base, improve performance, and fix error with FireFox. Added support for the dx and dy attributes in text elements for animations in the gSVG template engine. Updated the template engine. This allows for better control of text element placement in animations. In addition, added new test cases related to these changes