feat(tools): add map option to packer tool
The packer tool has been updated to include a new command line option, '-x/--map', to allow the option to build a .map file. Adjustments have been also made in the pack.js file to incorporate this option into the bundling process. If the option is not invoked, the tool follows the usual procedure of versioning, optimizing, and processing the files.