fix(build): fix plugin building issues
It fixes issues in the `package.json` where the build path for different plugins was incorrect due to a prefixed `gsvg.` which has now been removed. This commit updates the version and date format in the CHANGELOG.md file, along with adding recent bug fixes related to plugin loading and building.